+++ mammoth/utils/statistics.py
@@ -65,7 +65,7 @@ class Statistics(object):
             our_stats(list([`Statistics`])): list of updated stats
         """
         from torch.distributed import get_rank
-        from mammoth.utils.distributed import all_gather_list
+        from mammoth.distributed import all_gather_list
 
         # Get a list of world_size lists with len(stat_list) Statistics objects
         all_stats = all_gather_list(stat_list, max_size=max_size)
